# compiler(arm64): avoid conditional jumps to the trap handler