@ncruces 2023-06-19
Should I send a PR with the missing `return` on `amd64`?

I also owe you guys a TinyGo test to prove the multiple `panic` correct line numbers.