@evacchi 2023-06-19
> Should I send a PR with the missing `return` on `amd64`?

sure go ahead